build(deps): update dependencies including tree-sitter 0.26
Update all dependencies to latest compatible versions for security patches,
bug fixes, and performance improvements.

Key updates:
- tree-sitter: 0.25.6 → 0.26.3 (breaking: Node::child() now takes u32)
- tokio: 1.45.1 → 1.48.0
- serde: 1.0.219 → 1.0.228
- regex: 1.11.1 → 1.12.2
- reqwest: 0.12.25 → 0.12.27

Breaking change adaptation:
- refactor.rs:719: cast loop index to u32 for tree-sitter 0.26 API

Version constraints relaxed in Cargo.toml for easier future updates.
All 257 tests pass, clippy clean.

Sprint 46 / PBI-065

+++ src/analysis/refactor.rs
@@ -716,7 +716,7 @@ fn handle_code_actions_with_context(
 
     let child_count = params_node.child_count();
     for i in 0..child_count {
-        if let Some(ch) = params_node.child(i) {
+        if let Some(ch) = params_node.child(i as u32) {
             match ch.kind() {
                 "(" => {
                     lparen_end = Some(ch.end_byte());
